+++ ts-client/src/dlmm/index.ts
@@ -34,6 +34,7 @@ import {
   BIN_ARRAY_FEE_BN,
   DEFAULT_BIN_PER_POSITION,
   FEE_PRECISION,
+  FunctionType,
   MAX_ACTIVE_BIN_SLIPPAGE,
   MAX_BINS_PER_POSITION,
   MAX_BIN_ARRAY_SIZE,
@@ -1345,6 +1346,7 @@ export class DLMM {
       activeId: activeId.toNumber(),
       binStep: binStep.toNumber(),
       baseFactor: baseFactor.toNumber(),
+      functionType: FunctionType.LiquidityMining,
       activationType,
       activationPoint: activationPoint ? activationPoint : null,
       hasAlphaVault,
@@ -1355,6 +1357,8 @@ export class DLMM {
       padding: Array(63).fill(0),
     };
 
+    const preInstructions: TransactionInstruction[] = [];
+
     const userTokenX = getAssociatedTokenAddressSync(
       tokenX,
       creatorKey,
@@ -1362,12 +1366,66 @@ export class DLMM {
       tokenXAccount.owner
     );
 
+    const createUserTokenXIx =
+      createAssociatedTokenAccountIdempotentInstruction(
+        creatorKey,
+        userTokenX,
+        creatorKey,
+        tokenX,
+        tokenXAccount.owner
+      );
+
+    preInstructions.push(createUserTokenXIx);
+
     const userTokenY = getAssociatedTokenAddressSync(
       tokenY,
       creatorKey,
       true,
       tokenYAccount.owner
     );
+
+    const createUserTokenYIx =
+      createAssociatedTokenAccountIdempotentInstruction(
+        creatorKey,
+        userTokenY,
+        creatorKey,
+        tokenY,
+        tokenYAccount.owner
+      );
+
+    preInstructions.push(createUserTokenYIx);
+
+    const postInstructions: TransactionInstruction[] = [];
+
+    // if either mint (tokenX or tokenY) is SOL, wrap a small amount to initialize the wrapped SOL account(s) then unwrap after the pool creation
+    if (
+      (tokenX.equals(NATIVE_MINT) || tokenY.equals(NATIVE_MINT)) &&
+      !opt?.skipSolWrappingOperation
+    ) {
+      const wrapAmount = BigInt(1); // 1 lamport
+
+      if (tokenX.equals(NATIVE_MINT)) {
+        const wrapSOLIxX = wrapSOLInstruction(
+          creatorKey,
+          userTokenX,
+          wrapAmount
+        );
+        preInstructions.push(...wrapSOLIxX);
+      }
+      if (tokenY.equals(NATIVE_MINT)) {
+        const wrapSOLIxY = wrapSOLInstruction(
+          creatorKey,
+          userTokenY,
+          wrapAmount
+        );
+        preInstructions.push(...wrapSOLIxY);
+      }
+
+      const unwrapSOLIx = await unwrapSOLInstruction(creatorKey);
+      if (unwrapSOLIx) {
+        postInstructions.push(unwrapSOLIx);
+      }
+    }
 
     return program.methods
       .initializeCustomizablePermissionlessLbPair2(ixData)
@@ -1388,6 +1446,8 @@ export class DLMM {
         tokenProgramX: tokenXAccount.owner,
         tokenProgramY: tokenYAccount.owner,
       })
+      .preInstructions(preInstructions)
+      .postInstructions(postInstructions)
       .transaction();
   }
 
@@ -1428,6 +1488,9 @@ export class DLMM {
       program.programId
     );
 
+    const [tokenXAccount, tokenYAccount] =
+      await connection.getMultipleAccountsInfo([tokenX, tokenY]);
+
     const [reserveX] = deriveReserve(tokenX, lbPair, program.programId);
     const [reserveY] = deriveReserve(tokenY, lbPair, program.programId);
     const [oracle] = deriveOracle(lbPair, program.programId);
@@ -1452,6 +1515,7 @@ export class DLMM {
       activeId: activeId.toNumber(),
       binStep: binStep.toNumber(),
       baseFactor: baseFactor.toNumber(),
+      functionType: FunctionType.LiquidityMining,
       activationType,
       activationPoint: activationPoint ? activationPoint : null,
       hasAlphaVault,
@@ -1462,8 +1526,75 @@ export class DLMM {
       padding: Array(63).fill(0),
     };
 
-    const userTokenX = getAssociatedTokenAddressSync(tokenX, creatorKey);
-    const userTokenY = getAssociatedTokenAddressSync(tokenY, creatorKey);
+    const preInstructions: TransactionInstruction[] = [];
+
+    const userTokenX = getAssociatedTokenAddressSync(
+      tokenX,
+      creatorKey,
+      true,
+      tokenXAccount.owner
+    );
+
+    const createUserTokenXIx =
+      createAssociatedTokenAccountIdempotentInstruction(
+        creatorKey,
+        userTokenX,
+        creatorKey,
+        tokenX,
+        tokenXAccount.owner
+      );
+
+    preInstructions.push(createUserTokenXIx);
+
+    const userTokenY = getAssociatedTokenAddressSync(
+      tokenY,
+      creatorKey,
+      true,
+      tokenYAccount.owner
+    );
+
+    const createUserTokenYIx =
+      createAssociatedTokenAccountIdempotentInstruction(
+        creatorKey,
+        userTokenY,
+        creatorKey,
+        tokenY,
+        tokenYAccount.owner
+      );
+
+    preInstructions.push(createUserTokenYIx);
+
+    const postInstructions: TransactionInstruction[] = [];
+
+    // if either mint (tokenX or tokenY) is SOL, wrap a small amount to initialize the wrapped SOL account(s) then unwrap after the pool creation
+    if (
+      (tokenX.equals(NATIVE_MINT) || tokenY.equals(NATIVE_MINT)) &&
+      !opt?.skipSolWrappingOperation
+    ) {
+      const wrapAmount = BigInt(1); // 1 lamport
+
+      if (tokenX.equals(NATIVE_MINT)) {
+        const wrapSOLIxX = wrapSOLInstruction(
+          creatorKey,
+          userTokenX,
+          wrapAmount
+        );
+        preInstructions.push(...wrapSOLIxX);
+      }
+      if (tokenY.equals(NATIVE_MINT)) {
+        const wrapSOLIxY = wrapSOLInstruction(
+          creatorKey,
+          userTokenY,
+          wrapAmount
+        );
+        preInstructions.push(...wrapSOLIxY);
+      }
+
+      const unwrapSOLIx = await unwrapSOLInstruction(creatorKey);
+      if (unwrapSOLIx) {
+        postInstructions.push(unwrapSOLIx);
+      }
+    }
 
     return program.methods
       .initializeCustomizablePermissionlessLbPair(ixData)
@@ -1480,6 +1611,8 @@ export class DLMM {
         userTokenY,
         funder: creatorKey,
       })
+      .preInstructions(preInstructions)
+      .postInstructions(postInstructions)
       .transaction();
   }
 
@@ -1841,7 +1974,7 @@ export class DLMM {
       .setPairStatusPermissionless(status)
       .accountsPartial({
         lbPair: this.pubkey,
-        creator,
+        signer: creator,
       })
       .transaction();
 
@@ -2457,6 +2590,7 @@ export class DLMM {
         position: positionPubKey,
         lbPair: this.pubkey,
         owner: user,
+        rent: SYSVAR_RENT_PUBKEY,
       })
       .instruction();
 
@@ -2817,6 +2951,7 @@ export class DLMM {
           lbPair: this.pubkey,
           owner,
           payer,
+          rent: SYSVAR_RENT_PUBKEY,
         })
         .instruction();
 
@@ -2954,6 +3089,7 @@ export class DLMM {
         position: positionPubKey,
         lbPair: this.pubkey,
         owner: user,
+        rent: SYSVAR_RENT_PUBKEY,
       })
       .instruction();
     preInstructions.push(initializePositionIx);
@@ -3159,6 +3295,7 @@ export class DLMM {
         position: positionPubKey,
         lbPair: this.pubkey,
         owner: user,
+        rent: SYSVAR_RENT_PUBKEY,
       })
       .instruction();
     preInstructions.push(initializePositionIx);
@@ -5197,7 +5334,7 @@ export class DLMM {
       .setActivationPoint(activationPoint)
       .accountsPartial({
         lbPair: this.pubkey,
-        admin: this.lbPair.creator,
+        signer: this.lbPair.creator,
       })
       .transaction();
 
@@ -5217,7 +5354,7 @@ export class DLMM {
       .setPairStatus(pairStatus)
       .accountsPartial({
         lbPair: this.pubkey,
-        admin: this.lbPair.creator,
+        signer: this.lbPair.creator,
       })
       .transaction();
 
@@ -7269,6 +7406,7 @@ export class DLMM {
         position,
         lbPair: this.pubkey,
         owner: user,
+        rent: SYSVAR_RENT_PUBKEY,
       })
       .instruction();
 
@@ -7683,7 +7821,7 @@ export class DLMM {
               supply: bin.liquiditySupply,
               feeAmountXPerTokenStored: bin.feeAmountXPerTokenStored,
               feeAmountYPerTokenStored: bin.feeAmountYPerTokenStored,
-              rewardPerTokenStored: bin.rewardPerTokenStored,
+              rewardPerTokenStored: bin.functionBytes,
               price: pricePerLamport,
               version: binArray.version,
               pricePerToken: new Decimal(pricePerLamport)