+++ ts-client/src/index.ts
@@ -5,6 +5,6 @@ export * from "./dlmm/helpers";
 export * from "./dlmm/types";
 export * from "./dlmm/error";
 export * from "./dlmm/constants";
-export * from "./dlmm/idl";
-export { default as IDL } from "./dlmm/dlmm.json";
+export * from "./dlmm/idl/idl";
+export { default as IDL } from "./dlmm/idl/idl.json";
 export * from "./dlmm/helpers/accountFilters";