+++ ts-client/src/test/rebalance_with_strategy.test.ts
@@ -14,8 +14,12 @@ import {
   Transaction,
 } from "@solana/web3.js";
 import fs from "fs";
-import { BASIS_POINT_MAX, LBCLMM_PROGRAM_IDS } from "../dlmm/constants";
-import IDL from "../dlmm/dlmm.json";
+import {
+  BASIS_POINT_MAX,
+  FunctionType,
+  LBCLMM_PROGRAM_IDS,
+} from "../dlmm/constants";
+import IDL from "../dlmm/idl/idl.json";
 import {
   deriveLbPairWithPresetParamWithIndexKey,
   derivePresetParameterWithIndex,
@@ -28,7 +32,9 @@ import {
   assertionWithPercentageTolerance,
   assertionWithTolerance,
   createTestProgram,
+  createWhitelistOperator,
   logPositionLiquidities,
+  OperatorPermission,
   swap,
 } from "./helper";
 
@@ -216,15 +222,24 @@ describe("Rebalance with strategy", () => {
 
     const program = createTestProgram(connection, programId, keypair);
 
+    const operatorPda = await createWhitelistOperator(
+      connection,
+      keypair,
+      keypair.publicKey,
+      [OperatorPermission.InitializePresetParameter],
+      programId
+    );
+
     const presetParamState2 =
       await program.account.presetParameter.fetchNullable(presetParamPda2);
 
     if (!presetParamState2) {
       await program.methods
-        .initializePresetParameter2({
+        .initializePresetParameter({
           index: index.toNumber(),
           binStep: DEFAULT_BIN_STEP.toNumber(),
           baseFactor: DEFAULT_BASE_FACTOR_2.toNumber(),
+          functionType: FunctionType.LiquidityMining,
           filterPeriod: 30,
           decayPeriod: 600,
           reductionFactor: 5000,
@@ -234,9 +249,10 @@ describe("Rebalance with strategy", () => {
           baseFeePowerFactor: 0,
         })
         .accountsPartial({
-          admin: keypair.publicKey,
+          signer: keypair.publicKey,
           presetParameter: presetParamPda2,
           systemProgram: web3.SystemProgram.programId,
+          operator: operatorPda,
         })
         .signers([keypair])
         .rpc({